add mock lobster and a lobster tutorial
