+++ tests/conftest.py
@@ -64,7 +64,7 @@ def tmp_dir():
 
 @pytest.fixture(scope="session")
 def debug_mode():
-    return False
+    return True
 
 
 @pytest.fixture(scope="session")
