+++ tutorials/mock_lobster.py
@@ -1,4 +1,4 @@
-"""Mock VASP functions for executing tutorials."""
+"""Mock LOBSTER functions for executing tutorials."""
 
 import contextlib
 import os
@@ -9,15 +9,15 @@ from pathlib import Path
 
 from pytest import MonkeyPatch
 
-from atomate2.utils.testing.vasp import monkeypatch_vasp
+from atomate2.utils.testing.lobster import monkeypatch_lobster
 
 TEST_ROOT = Path(__file__).parent.parent / "tests"
 TEST_DIR = TEST_ROOT / "test_data"
 
 
 @contextlib.contextmanager
-def mock_vasp(ref_paths: dict) -> Generator:
-    """Mock VASP functions.
+def mock_lobster(ref_paths: dict) -> Generator:
+    """Mock LOBSTER functions.
 
     Parameters
     ----------
@@ -27,14 +27,14 @@ def mock_vasp(ref_paths: dict) -> Generator:
     ------
         function: A function that mocks calls to VASP.
     """
-    for mf in monkeypatch_vasp(MonkeyPatch(), TEST_DIR / "vasp"):
-        fake_run_vasp_kwargs = {k: {"check_inputs": ()} for k in ref_paths}
+    for mf in monkeypatch_lobster(MonkeyPatch(), TEST_DIR / "lobster"):
+        fake_run_lobster_kwargs = {k: {"lobsterin_settings": ()} for k in ref_paths}
         old_cwd = os.getcwd()
         new_path = tempfile.mkdtemp()
         os.chdir(new_path)
         try:
-            yield mf(ref_paths, fake_run_vasp_kwargs)
+            yield mf(ref_paths, fake_run_lobster_kwargs)
         finally:
             os.chdir(old_cwd)
-            shutil.rmtree(new_path)
+            #shutil.rmtree(new_path)
